Fix vint4.set_lane on NEON

+++ Source/astcenc_vecmathlib_neon_4.h
@@ -247,7 +247,7 @@ struct vint4
 	/**
 	 * @brief Set the scalar value of a single lane.
 	 */
-	template <int l> ASTCENC_SIMD_INLINE void set_lane(float a)
+	template <int l> ASTCENC_SIMD_INLINE void set_lane(int a)
 	{
 		m = vld1q_lane_s32(&a, m, l);
 	}
